fix(server): extract to hleper
